# [Refactor] Improve the performance of temporal group averaging