## Description
<!--
  Please include a summary of the change and which issue is fixed.
  Please also include relevant motivation and context.
  List any dependencies that are required for this change.
-->

- Closes #688 

## TODO:

- [x] Identify performance bottlenecks
   1. Generating labeled time coordinates (aka assign groups) and adding it to the existing time dimension with existing coordinates, then performing the Xarray groupby yields extremely slow results (not sure why, it's an Xarray issue). (Refer to [comment](https://github.com/xCDAT/xcdat/issues/688#issuecomment-2319299638)) -- replace time coords with labeled time coords directly for grouping, rather than adding labeled time coords as auxiliary coords on the time dimension (which slows things down in Xarray for some reason, need to ask Xarray forum)
   2. In `_get_weights()`, loading time lengths into memory is slow ([lines](https://github.com/xCDAT/xcdat/blob/584fccee91f559089e4bb24c9cc45983b998bf4a/xcdat/temporal.py#L1281-L1290)) -- replace with casting to `"timedelta64[ns]"` then `float64`
   3. In `_get_weights()`, performing validation to check the sums of weights for each group adds up to 1 is slow ([lines](https://github.com/xCDAT/xcdat/blob/584fccee91f559089e4bb24c9cc45983b998bf4a/xcdat/temporal.py#L1296-L1299)) -- remove this unnecessary assertion
- [ ] ~~Identify performance optimizations -- I don't think this is necessary right now~~
   1. Xarray `groupby` with vs. without `flox` package
   2. Try with Dask chunking
- [x] Make sure unit tests still pass
- [x] Measure performance difference between this branch and `main`
- [x] Perform regression testing between branch code on same dataset

## Checklist

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] My changes generate no new warnings
- [ ] Any dependent changes have been merged and published in downstream modules

If applicable:

- [ ] I have added tests that prove my fix is effective or that my feature works
- [x] New and existing unit tests pass with my changes (locally and CI/CD build)
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [ ] I have noted that this is a breaking change for a major release (fix or feature that would cause existing functionality to not work as expected)